Try to fix command line parsing issues with UNIX style absolute paths.
Previously a path like /home/user/myboogie_program.bpl would be
considered a flag so it wasn't actually possible to use an absolute
path under Linux/OSX to corral.exe

This is certainly not perfect. Using '/' as the prefix for command line
parameters is a really bad idea exactly for the reason described above.