Fix command line argument parsing of the input file. Previously under operating systems with UNIX file separators (e.g. "/path/to/myfile.bpl") the path would be interpreted as a command flag making the tool unusable.
This is fixed by enforcing that the first argument (as stated in the usage
message) to the tool is always the input bpl file which is never
considered to be a flag. All other arguments are considered to be flags.

I would prefer if '/' was not used as the prefix for flag characters but
doing so would break the current clients of Corral.